backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pydantic
python-dotenv
python-docx
certifi
scikit-learn>=1.3.0
numpy>=1.24.0
